generate models and client

+++ client/model_cash_letter_control.go
@@ -30,5 +30,5 @@ type CashLetterControl struct {
 	// SettlementDate identifies the date that the institution that creates the cash letter expects settlement.
 	SettlementDate time.Time `json:"settlementDate"`
 	// CreditTotalIndicator is a code that indicates whether Credits Items are included in this record’s totals. If so they will be included in TotalItemCount and FileTotalAmount. TotalRecordCount includes all records of all types regardless of the value of this field. * ` ` - No Credit Items * `0` - Credit Items are not included in totals * `1` - Credit Items are included in totals
-	CreditTotalIndicator float32 `json:"creditTotalIndicator,omitempty"`
+	CreditTotalIndicator int32 `json:"creditTotalIndicator,omitempty"`
 }
